Make serializer interface more obvious
